requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
Flask-Limiter==3.5.0
Pillow>=9.0.0
gunicorn==21.2.0 